added product options for the user to select
added options to the product form, checkout view, checkout model, cart view and contexts
added css for the options displayed on the product detail page.
currently the user cannot add a duplicate product with different options to the cart
because it is a dictionairy, due to time constraints this will be fixed in a futire release,
for now the user will get an error message explaining why they cannot add a duplicate.